refactor: rename maintenance time-check utility
