+++ docs/20260206_1056_メンテナンスモード運用手順.md
@@ -5,7 +5,7 @@
 - API (`/api/*`) の停止/再開
 
 ## 現在の実装仕様
-- 判定ロジック: `src/lib/utils/time-check.ts`
+- 判定ロジック: `src/lib/utils/maintenance-mode.ts`
 - 開始時刻定数: `MAINTENANCE_START_AT`
 - 事前確認フラグ: `?preview=maintenance`
 - メンテ中の挙動:
@@ -13,29 +13,29 @@
   - API: `503` (`Retry-After` 付き)
 
 ## ON（メンテナンス開始）オペレーション
-1. `src/lib/utils/time-check.ts` の `MAINTENANCE_START_AT` を開始したい JST 時刻に設定
+1. `src/lib/utils/maintenance-mode.ts` の `MAINTENANCE_START_AT` を開始したい JST 時刻に設定
 - 例: `new Date("2026-02-07T23:59:00+09:00")`
 2. 反映前の見た目確認
 - `/?preview=maintenance` にアクセスして停止画面を確認
 3. テスト実行
-- `pnpm run test:unit -- src/lib/utils/time-check.test.ts`
+- `pnpm run test:unit -- src/lib/utils/maintenance-mode.test.ts`
 4. コミット・デプロイ
 
 ## OFF（メンテナンス終了）オペレーション
-1. `src/lib/utils/time-check.ts` の `MAINTENANCE_START_AT` を `null` に更新
+1. `src/lib/utils/maintenance-mode.ts` の `MAINTENANCE_START_AT` を `null` に更新
 - 例: `export const MAINTENANCE_START_AT: Date | null = null;`
 2. 動作確認
 - `/maintenance` が `/` へリダイレクトされること
 - 通常ページが表示されること
 - `/api/*` が `503` ではなく通常応答になること
 3. テスト実行
-- `pnpm run test:unit -- src/lib/utils/time-check.test.ts`
+- `pnpm run test:unit -- src/lib/utils/maintenance-mode.test.ts`
 4. コミット・デプロイ
 
 ## リリース前チェックリスト
 - [ ] ON時は開始日時を JST (`+09:00`) で指定した / OFF時は `null` にした
 - [ ] `/?preview=maintenance` で停止画面確認済み
-- [ ] `pnpm run test:unit -- src/lib/utils/time-check.test.ts` が PASS
+- [ ] `pnpm run test:unit -- src/lib/utils/maintenance-mode.test.ts` が PASS
 - [ ] API の `503` / 通常応答を期待どおり確認した
 
 ## 備考
